requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argcomplete==3.6.2
boto3==1.37.33
botocore==1.37.33
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.2
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
fastavro==1.10.0
filelock==3.18.0
firecrawl-py==1.15.0
fsspec==2025.3.2
google-auth==2.38.0
griffe==1.7.2
groq==0.22.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.51
langgraph==0.3.29
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.30
lightrag-hku==1.3.1
logfire-api==3.14.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
mistralai==1.6.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.72.0
opentelemetry-api==1.32.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-ai==0.0.55
pydantic-ai-slim==0.0.55
pydantic-evals==0.0.55
pydantic-graph==0.0.55
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-frontmatter==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tenacity==9.1.2
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
zipp==3.21.0
zstandard==0.23.0
tavily-python